fix bug
修正了分页时参数错误导致数据重复的问题

+++ LookNovel/classes/base/controller/LNTableViewController.h
@@ -27,7 +27,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)setDataComplete;
 
-- (float)pageSize;
+- (NSInteger)pageSize;
 
 #pragma mark -
 //下拉刷新
